% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\docType{class}
\name{geocode-class}
\alias{geocode-class}
\alias{geocode}
\title{Coordinate finder}
\description{
The 'geocode' contains a function that converts an address to coordinates using an openweather API.
}
\section{Fields}{

\describe{
\item{\code{out}}{A data frame contains address, lontitude, latitude.}

\item{\code{adrs}}{A character represents address}
}}


