% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5G26.R
\docType{package}
\name{lab5G26}
\alias{lab5G26}
\alias{lab5G26-package}
\title{LAB5}
\description{
This package LAB5G26 contains a function using API for map searching.
}
\author{
Mubarak Hussain
}
